+++ stellar-dotnet-sdk/requests/AssetsRequestBuilder.cs
@@ -27,7 +27,7 @@ namespace stellar_dotnet_sdk.requests
         /// <returns></returns>
         public AssetsRequestBuilder AssetCode(string assetCode)
         {
-            _uriBuilder.SetQueryParam("asset_code", assetCode);
+            UriBuilder.SetQueryParam("asset_code", assetCode);
             return this;
         }
 
@@ -38,7 +38,7 @@ namespace stellar_dotnet_sdk.requests
         /// <returns></returns>
         public AssetsRequestBuilder AssetIssuer(string assetIssuer)
         {
-            _uriBuilder.SetQueryParam("asset_issuer", assetIssuer);
+            UriBuilder.SetQueryParam("asset_issuer", assetIssuer);
             return this;
         }
     }
