+++ stellar-dotnet-sdk/requests/TradesRequestBuilder.cs
@@ -18,11 +18,11 @@ namespace stellar_dotnet_sdk.requests
 
         public TradesRequestBuilder BaseAsset(Asset asset)
         {
-            _uriBuilder.SetQueryParam("base_asset_type", asset.GetType());
+            UriBuilder.SetQueryParam("base_asset_type", asset.GetType());
             if (asset is AssetTypeCreditAlphaNum creditAlphaNumAsset)
             {
-                _uriBuilder.SetQueryParam("base_asset_code", creditAlphaNumAsset.Code);
-                _uriBuilder.SetQueryParam("base_asset_issuer", creditAlphaNumAsset.Issuer.AccountId);
+                UriBuilder.SetQueryParam("base_asset_code", creditAlphaNumAsset.Code);
+                UriBuilder.SetQueryParam("base_asset_issuer", creditAlphaNumAsset.Issuer.AccountId);
             }
 
             return this;
@@ -30,17 +30,17 @@ namespace stellar_dotnet_sdk.requests
 
         public TradesRequestBuilder OfferId(string offerId)
         {
-            _uriBuilder.SetQueryParam("offer_id", offerId);
+            UriBuilder.SetQueryParam("offer_id", offerId);
             return this;
         }
 
         public TradesRequestBuilder CounterAsset(Asset asset)
         {
-            _uriBuilder.SetQueryParam("counter_asset_type", asset.GetType());
+            UriBuilder.SetQueryParam("counter_asset_type", asset.GetType());
             if (asset is AssetTypeCreditAlphaNum creditAlphaNumAsset)
             {
-                _uriBuilder.SetQueryParam("counter_asset_code", creditAlphaNumAsset.Code);
-                _uriBuilder.SetQueryParam("counter_asset_issuer", creditAlphaNumAsset.Issuer.AccountId);
+                UriBuilder.SetQueryParam("counter_asset_code", creditAlphaNumAsset.Code);
+                UriBuilder.SetQueryParam("counter_asset_issuer", creditAlphaNumAsset.Issuer.AccountId);
             }
 
             return this;
