+++ stellar-dotnet-sdk/responses/AccountResponse.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace stellar_dotnet_sdk.responses
@@ -61,129 +60,5 @@ namespace stellar_dotnet_sdk.responses
         {
             SequenceNumber++;
         }
-    }
-
-    /// <summary>
-    ///     Represents account thresholds.
-    /// </summary>
-    public class Thresholds
-    {
-        public Thresholds(int lowThreshold, int medThreshold, int highThreshold)
-        {
-            LowThreshold = lowThreshold;
-            MedThreshold = medThreshold;
-            HighThreshold = highThreshold;
-        }
-
-        [JsonProperty(PropertyName = "low_threshold")]
-        public int LowThreshold { get; private set; }
-
-        [JsonProperty(PropertyName = "med_threshold")]
-        public int MedThreshold { get; private set; }
-
-        [JsonProperty(PropertyName = "high_threshold")]
-        public int HighThreshold { get; private set; }
-    }
-
-    /// <summary>
-    ///     Represents account flags.
-    /// </summary>
-    public class Flags
-    {
-        public Flags(bool authRequired, bool authRevocable)
-        {
-            AuthRequired = authRequired;
-            AuthRevocable = authRevocable;
-        }
-
-        [JsonProperty(PropertyName = "auth_required")]
-        public bool AuthRequired { get; private set; }
-
-        [JsonProperty(PropertyName = "auth_revocable")]
-        public bool AuthRevocable { get; private set; }
-    }
-
-    /// <summary>
-    ///     Represents account balance.
-    /// </summary>
-    public class Balance
-    {
-        [JsonProperty(PropertyName = "asset_issuer")]
-        private string _assetIssuer;
-
-        public Balance(string assetType, string assetCode, string assetIssuer, string balance, string limit, string buyingLiabilities, string sellingLiabilities)
-        {
-            AssetType = assetType ?? throw new ArgumentNullException(nameof(assetType), "assertType cannot be null");
-            BalanceString = balance ?? throw new ArgumentNullException(nameof(balance), "balance cannot be null");
-            Limit = limit;
-            AssetCode = assetCode;
-            _assetIssuer = assetIssuer;
-            BuyingLiabilities = buyingLiabilities ?? throw new ArgumentNullException(nameof(buyingLiabilities), "buyingLiabilities cannot be null");
-            SellingLiabilities = sellingLiabilities ?? throw new ArgumentNullException(nameof(sellingLiabilities), "sellingLiabilities cannot be null");
-        }
-
-        [JsonProperty(PropertyName = "asset_type")]
-        public string AssetType { get; private set; }
-
-        [JsonProperty(PropertyName = "asset_code")]
-        public string AssetCode { get; private set; }
-
-        //This prop is dynamic based on private field serialized above.
-        public KeyPair AssetIssuer => KeyPair.FromAccountId(_assetIssuer);
-
-        [JsonProperty(PropertyName = "limit")] public string Limit { get; private set; }
-
-        [JsonProperty(PropertyName = "balance")]
-        public string BalanceString { get; private set; }
-
-        [JsonProperty(PropertyName = "buying_liabilities")]
-        public string BuyingLiabilities { get; private set; }
-
-        [JsonProperty(PropertyName = "selling_liabilities")]
-        public string SellingLiabilities { get; private set; }
-    }
-
-    /// <summary>
-    ///     Represents account signers.
-    /// </summary>
-    public class Signer
-    {
-        public Signer(string accountId, int? weight)
-        {
-            AccountId = accountId ?? throw new ArgumentNullException(nameof(accountId), "accountId cannot be null");
-            Weight = weight ?? throw new ArgumentNullException(nameof(weight), "weight cannot be null");
-        }
-
-        [JsonProperty(PropertyName = "public_key")]
-        public string AccountId { get; private set; }
-
-        [JsonProperty(PropertyName = "weight")]
-        public int Weight { get; private set; }
-    }
-
-    public class AccountResponseLinks
-    {
-        public AccountResponseLinks(Link effects, Link offers, Link operations, Link self, Link transactions)
-        {
-            Effects = effects;
-            Offers = offers;
-            Operations = operations;
-            Self = self;
-            Transactions = transactions;
-        }
-
-        [JsonProperty(PropertyName = "effects")]
-        public Link Effects { get; private set; }
-
-        [JsonProperty(PropertyName = "offers")]
-        public Link Offers { get; private set; }
-
-        [JsonProperty(PropertyName = "operations")]
-        public Link Operations { get; private set; }
-
-        [JsonProperty(PropertyName = "self")] public Link Self { get; private set; }
-
-        [JsonProperty(PropertyName = "transactions")]
-        public Link Transactions { get; private set; }
     }
 }
