+++ stellar-dotnet-sdk/responses/Flags.cs
@@ -1,0 +1,22 @@
+﻿using Newtonsoft.Json;
+
+namespace stellar_dotnet_sdk.responses
+{
+    /// <summary>
+    ///     Represents account flags.
+    /// </summary>
+    public class Flags
+    {
+        public Flags(bool authRequired, bool authRevocable)
+        {
+            AuthRequired = authRequired;
+            AuthRevocable = authRevocable;
+        }
+
+        [JsonProperty(PropertyName = "auth_required")]
+        public bool AuthRequired { get; private set; }
+
+        [JsonProperty(PropertyName = "auth_revocable")]
+        public bool AuthRevocable { get; private set; }
+    }
+}
